Bugfix: Wrong data type had been used for TUBESIZE
This results in an overflow or the argument to malloc() on large
screens.

Some uninitialized variable issues fixed.

+++ ex_vis.h
@@ -104,7 +104,7 @@ var enum {
  */
 var	short	TUBELINES;
 var	short	TUBECOLS;
-var	short	TUBESIZE;
+var	int	TUBESIZE;
 
 /*
  * The screen in visual and crtopen is of varying size; the basic
